오류수정 (#13)
* fix : 값이 존재하지 않을때 404띄우기

* fix : 404 띄우기 코드 추가

+++ src/main/java/com/poppin/backendserver/controller/ReviewController.java
@@ -53,6 +53,10 @@ public class ReviewController {
      */
     @PutMapping("/{review_id}")
     public ResponseEntity<Long> updateReview(@PathVariable("review_id") Long id, @RequestBody ReviewDTO reviewDTO) {
+        Optional<Review> savedReview = reviewService.getReview(id);
+        if (savedReview.isPresent()) {
+            return ResponseEntity.notFound().build();
+        }
         Review review = Review.builder()
                 .rating(reviewDTO.getRating())
                 .title(reviewDTO.getTitle())
@@ -67,6 +71,10 @@ public class ReviewController {
      */
     @DeleteMapping("/{review_id}")
     public ResponseEntity<Long> deleteReview(@PathVariable("review_id") Long id) {
+        Optional<Review> savedReview = reviewService.getReview(id);
+        if (savedReview.isPresent()) {
+            return ResponseEntity.notFound().build();
+        }
         Long deletedReviewId = reviewService.deleteReview(id);
         return ResponseEntity.ok(deletedReviewId);
     }
